Update patch set 5
Patch Set 5: Commit-Queue+1

> Patch Set 4:
> 
> Sorry about that.

No worries!

> IMO it would be better to have `RSAZ_1024_mod_exp_avx2` return `powerBufLen` so that the OPENSSL_cleanse call in exponentiation.c would do the correct thing.

Hrm. That's a thought. It's not a pattern we've used elsewhere, but there is something to be said for moving cleansing closer to when the buffer is actually deallocated.

Arguably cleansing stack-allocated should be the purview of the compiler, with some build flag or function attribute or something to get it to just clear it for us. The compiler is the only one that knows about register spills and such.

To that end, maybe that entire buffer should be cleansed, rather than thinking hard about powerbufLen...

(Currently just trying to deal with CQ failures from upper ones in the chain.)

Patch-set: 5
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Commit-Queue=+1