Update patch set 8
Patch Set 8:

> Patch Set 6:
> > Hrm. That's a thought. It's not a pattern we've used elsewhere, but there is something to be said for moving cleansing closer to when the buffer is actually deallocated.

Yes, that was my thinking: allocation, deallocation, and zeroing should be done in the same function.

> > Arguably cleansing stack-allocated should be the purview of the compiler

I think "Arguably" is understating it; it definitely should be done automatically by the compiler.

> Hrm. It's a little fussy right now since the RSAZ path doesn't actually set up powerbuf at all and there's all the ifdefs around. *shrug* This works, I guess. Maybe someday compilers will get fancy stack-clearing options and we can put all this where it really should go.

It isn't wrong. But, note that the RSAZ case has two calls to OPENSSL_cleanse(), one of which does nothing. A minor thing, of course.

Patch-set: 8