Update patch set 1
Patch Set 1:

> It's mostly servers pushing way too many CBC IVs that care.
> This means bsaes's current side channel tradeoffs make even
> less sense here.

NIST says "There are two recommended methods for generating unpredictable IVs. The first method is to apply the forward cipher function, under the same key that is used for the encryption of the plaintext, to a nonce. The nonce must be a data block that is unique to each execution of the encryption operation. For example, the nonce may be a counter, as described in Appendix B, or a message number. The second method is to generate a random data block using a FIPS approved
random number generator."

Using an encrypted counter may solve this and other performance issues.

Patch-set: 1
CC: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>