Harden against fork via MADV_WIPEONFORK.
Linux 4.14 and up support MADV_WIPEONFORK, which can be used to reliably and
efficiently detect forks. It seemed like fun. Use it to harden the PRNG and RSA
blinding caches.  Now it takes quite a lot of work to get a fork-unsafe PRNG
out of BoringSSL on Linux. You must do all of:

- Explicitly disavow fork safety with RAND_enable_fork_unsafe_buffering
- Lack RDRAND
- Be running a kernel older than 4.14

On the RSA side, we currently cache blinding values. (Alas, removing the cache
gives a *very* noticeable perf hit. There is some low-hanging fruit to trim a
few Montgomery reductions, but it didn't offset much last I toyed with it.)
Now, on Linux 4.14 and up, this cache is fork-safe.

Since not all platforms that fork support fork detection, this should only be
used as a hardening measure. However, if it returns anything, I believe the
fork detection is reliable so we could, in theory, gate some optimizations on
this. (This CL does not do so.)

Upgrade-Note: BoringSSL now calls some more syscalls on Linux. If this offends
sandboxes, let us know. We can loosen the sandbox or add a mechanism to prime
the MADV_WIPEONFORK page before entering it.

Change-Id: I6ba43951aeaa2b9b81f74f9e5a7a0ce2de0438a4