Update patch set 3
Patch Set 3:

> ecp_nistz256_ord_sqr_mont. The rep parameter is typed int, but the code expected uint64_t.

Revisiting this now, I think it would be better to have the `rep` parameter be `crypto_word_t` or `BN_ULONG` or whatever type you are using to mean "full word". If somebody were to port this function to 32-bit platforms then `uint64_t` would be the wrong type.

Patch-set: 3