Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > ecp_nistz256_ord_sqr_mont. The rep parameter is typed int, but the code expected uint64_t.
> > 
> > Revisiting this now, I think it would be better to have the `rep` parameter be `crypto_word_t` or `BN_ULONG` or whatever type you are using to mean "full word". If somebody were to port this function to 32-bit platforms then `uint64_t` would be the wrong type.
> 
> That's what this CL did. :-P

(I'm guessing you only looked at the assembly and not the C code. A lot of the assembly comments seem to "compile out" BN_ULONG to uint64_t or uint32_t as appropriate, including this one, whereas the C ones use the general type. No particular opinions as to whether that's the best or not, but I was just continuing that practice.)

Patch-set: 3