Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Sorry for the delay, I've been traveling.
> 
> These terrible APIs make me cringe, but we do want to support Node. Before I take a deep breath and submit this: is there any way that we could sync these things to the intersection, not the union? I.e. remove as much as possible? (Because, seriously, output feedback mode?!)

I'm with you on not wanting to have ciphers that are awful decisions for any actual code exposed here, but the reality is if I remove them here I'll end up re-adding them as a patch in Electron. Electron (and node) are in a bit of an odd place here, because often people who want to use these terrible ciphers don't have a choice: for instance, they're communicating with a medical device[1] that has that cipher built in, or implementing some cryptocurrency nonsense[2].

That said, I haven't seen any requests for the -ofb ciphers yet, so maybe we can get away with removing them and hoping nobody asks? 😬🤞

[1]: https://github.com/electron/electron/issues/16195#issuecomment-455387971
[2]: https://github.com/electron/electron/issues/16195#issuecomment-455169309

Patch-set: 4
Reviewer: Gerrit User 9244 <9244@66b711a4-aa9b-3dac-9062-916540830212>