Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Any idea why only the ARMv8 code has a 2x variant? In particular, why doesn't x86-64 have one?

Just speculating, but ARMv8 has a *lot* of SIMD registers (32 of them), so it's easy run two of them in parallel. x86_64 only has 16. Then again, x86 only has 8 and vpaes-x86.pl seems to manage...

Also it seems vpaes doesn't perform all that great on ARMv8, so I guess this was tried as an optimization?

Patch-set: 3