Update patch set 9
Patch Set 9:

Code aligned with comments.
I've compiled it with GCC 4.8.4 and NDK r18 for aarch64 and armv7. Tests seem to work OK on x86 and hikey970.
I'm didn't had a chance to compile for macos and windows. Is it possible to run CI to see results?

Patch-set: 9
Reviewer: Gerrit User 9225 <9225@66b711a4-aa9b-3dac-9062-916540830212>