Update patch set 2
Patch Set 2: Code-Review+1

Sorry was confused with the `Ack` and `Done` in the comments.
I updated the change as per your comments David. Also, thanks a lot for the comments about getting the min and max versions. That certainly sounds reasonable. I work on electron and not directly on node, but will definitely take a look at how they implement it, or will convey this to folks.

Patch-set: 2
Reviewer: Gerrit User 9463 <9463@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+1