Update patch set 16
Patch Set 16:

> Patch Set 16:
> 
> > Patch Set 16: Commit-Queue+1
> > 
> > (Just going to watch the Windows builders blow up since I won't have access to a Windows machine for a few days. It's unfortunate that Windows builds w/o assembly support and I'll stare at the asm and error a little first.)
> 
> I can do assembly in windows if needed. I havn't done because I didn't had an access to windows machine. Original code from Microsoft also doesn't use assembly in windows, but I'm not sure why it was done this way.
> https://github.com/Microsoft/PQCrypto-SIDH/blob/b3de079aa4f82c41939eb4573fc9d3e73b899c42/src/P503/AMD64/fp_x64.c#L21
> 
> But that's not a problem at all - I can setup windows machine and make it build.

Especially if we're doing Windows, I think this should be probably get translated to perlasm. That way it can use the perlasm SEH directives when I finally finish that work. (We use NASM, not MASM, on Windows. NASM doesn't have support for unwind directives, so it'll end up implemented in the perlasm driver.) In the meantime, ghash-ssse3-x86_64.pl and trampoline-x86_64.pl have examples of doing unwind codes by hand, and the CHECK_ABI_SEH ABI testing variant will test it for correctness.

In particular, unwind codes are necessary for Chrome's sampling profiler to pick the function up, so we know if the CPU overhead is too much. (Although we still need to get the sampling profiler turned back on for the net stack after it got moved out-of-process...)

Note also Windows treats some xmm registers as callee-saved. The ABI tester will test this, and we should be adding ABI tests for new assembly anyway.

Patch-set: 16
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>