Don't try reading forever in bssl_shim.
This is bizarre, but with kernel 5.0.7 about five to ten runner tests
will fail in any given run because they're waiting to drain a socket.

bssl_shim drains sockets before closing them to avoid write errors in
runner. But, with this kernel, that's hanging. The socket on the shim
side is in FIN_WAIT2 (i.e. remote side has ACKed the FIN, but hasn't
sent their own FIN). On the runner side, the socket has been closed.
That suggests that FIN packets are getting lost over loopback, although
the counters say no loss occurred.

Change-Id: Id1c8dc0178e46d431beb3783a751819d3b2a6b21