Disable the common name fallback on *any* SAN list.
This aligns with the Go crypto/x509 behavior and reduces the cases when the SAN
to CN fallback occurs. If the certificate is new enough to have a SAN list,
even if it only contains email or IP addresses, it is reasonable to assume it
new enough that the common name is not a DNS name.

Update-Note: Our certificate verification is getting slightly stricter.
Change-Id: I9e3466d8dd8a722405c546181a589f797efa43f9