Update patch set 1
Patch Set 1:

We don't appear to have a CLA on file for mrostecki@opensuse.org. See the links here:
https://boringssl.googlesource.com/boringssl/+/refs/heads/master/CONTRIBUTING.md

Going purely by your commit description, this seems odd. We build for aarch64 on both Android and iOS and these have been just fine. I would expect the compiler to figure it out. It can always make a largely stack-allocation and then bitwise AND the pointer. Is this perhaps a limitation in GCC?

Patch-set: 1
CC: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>