Update patch set 20
Patch Set 20:

> Patch Set 20:
> 
> p.s. are you interested in landing this, or are you planning on testing the smaller parameters?

I think I'm interested in landing this code.
We would like to compare SIKE against HRSS (or SI vs SL). The interesting part of SIKE is a smaller size of a public key. But using smaller parameters, will reduce public key from 378 to 330 bytes, which I think isn't a huge difference. There is a bigger difference in speed, but we know it is going to be around 30% faster, as this result is related to field size.
From the other hand - I've looked at differences between Round1 and Round2 submission. It isn't complicated to update existing code to smaller field, but there is some work that needs to be done and I'm not exactly sure how much time it will take (for example code for aarch64 needs to be written from scratch).
So, I'm very interested in implementing smaller field. But if we could land this code and update parameters later on then it would be ideal, as it would allow us to start collecting data and analyze it.

Patch-set: 20
Reviewer: Gerrit User 9225 <9225@66b711a4-aa9b-3dac-9062-916540830212>