Update patch set 2
Patch Set 2:

sdallocx is some alternate free that knows the size. Some folks in the internal repo wanted us to use it. Given that Windows malloc typically does typically provide it, and weak symbols don't exist on Windows anyway, it sounds like we should just not bother trying to detect it on Windows (i.e. use OPENSSL_WINDOWS as the define rather than _MSC_VER and just treat this was a persistent thing).

Apparently Abseil is defining sdallocx to free as a fallback, like us. However, when it doesn't have weak symbols available, it defines the fallback as a non-weak symbol! That seems nonsense and should be fixed in Abseil. It breaks in, well, exactly this scenario. (Only we're defining a non-weak fallback on accident and they're doing it intentionally.) It doesn't work if two libraries both try to do this, and Abseil doesn't own the sdallocx name, the malloc implementation does.

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>