Update patch set 1
Patch Set 1: Commit-Queue+1

Paperwork: CLA confirmed.

This has been on the list for a while. Perhaps we trust the SDE builders to test this, although David may be aware of issues in the code and he's out for a week or so.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Commit-Queue=+1