Support PKCS#8 blobs using PBES2 with HMAC-SHA256.
OpenSSL now defaults to HMAC-SHA256 instead of HMAC-SHA1. This is needed for
some of cryptography.io's tests to pass. For now, this does not add support for
creating such things or change the default, but I've added a TODO marking the
place where OpenSSL sticks the parameter in its API.

Change-Id: I29de2ee4fdcfa2363f94189f75930f55d5d0cf67