Discuss the doubling case in windowed Booth representation.
Using signed digits for point multiplication halves the size of the table,
which is a valuable optimization. However, it is dependent on the curve whether
the doubling case in point addition is reachable. Assuming my reasoning is
valid, the condition is:

  The non-trivial doubling case in single-point scalar multiplication may occur
  if and only the 2^(w-1) bit of the group order is zero. If it is one, this
  case may be ignored.

It would be nice to transcribe this to Coq someday but, for now, check in a
proof in prose.

This file is now 91% comments by line count.

Change-Id: I29b394289793db957f99e80734e10ed59a96fcec