Update patch set 3
Patch Set 3:

So this change is hanging around. Sounds like this code needs a bunch of work to get the annotations correct before enabling, which is fiddly. However, Gerrit isn't a TODO list and I don't want to leave this open forever. (We will likely get to it in time even if closed.)

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>