Update patch set 1
Patch Set 1: Code-Review+1

LGTM, happy that this seems to be helping.

It is worth noting that this will regress the quality of code compared to emitting a `cmov`. =/ An unfortunate tradeoff, but difficult to avoid given the current architecture of Clang and LLVM in particular.

Patch-set: 1
Reviewer: Gerrit User 9603 <9603@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+1